requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
lxml==4.1.1
PyYAML==3.12
request==0.0.22
requests==2.18.4
urllib3==1.22
